Add testing for the action_decider_helper method

+++ app/helpers/action_decider_helper.rb
@@ -1,6 +1,6 @@
 module ActionDeciderHelper
   def like_action_decider(post, current_user)
-    if post.likes.find_by(author_id: current_user.id)
+    if post.respond_to?(:likes) && post.likes.find_by(author_id: current_user.id)
       {
         method: 'delete',
         url: user_post_like_path(user_id: post.author.id, post_id: post.id),
